variable types upgrade

+++ ft_strchr.c
@@ -6,13 +6,14 @@
 /*   By: rexposit <rexposit@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/19 18:05:24 by rexposit          #+#    #+#             */
-/*   Updated: 2024/09/19 19:19:08 by rexposit         ###   ########.fr       */
+/*   Updated: 2024/09/20 11:33:35 by rexposit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>  //SUSTITUIR LIBRERIA CON LIBFT.H
 char	*ft_strchr(const char *s, int c)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
